requirements.txt
beautifulsoup4
lxml
openai-whisper
requests
yt_dlp

